+++ R/MetaboLightsParam.R
@@ -77,6 +77,7 @@
 #' @importClassesFrom ProtGenerics Param
 #'
 #' @examples
+#'
 #' library(MsExperiment)
 #' # Load a study with the mtblsId "MTBLS39" and selecting specific file pattern
 #' # as well as removing ontology and protocol information in the metadata.